requirements.txt
# Core Playwright and testing dependencies
playwright>=1.41.0
pytest>=7.4.0
pytest-playwright>=0.4.0

# Test reporting and utilities
pytest-html>=4.0.0
pytest-xdist>=3.0.0
python-dotenv>=1.0.0

# Optional: Allure reporting (comment out if not needed)
# allure-pytest>=2.13.0

# Alternative: Use pre-built wheels for problematic packages
# If you encounter build issues, try installing from conda-forge:
# conda install -c conda-forge playwright pytest
